MAT-6223: Fix minor visual bug with paper, refactor popover to single component
